added component to view all activities for a date
